Added user restrictions on room update/delete features

+++ FriendCircle/base/views.py
@@ -26,6 +26,8 @@ from django.contrib.auth.decorators import login_required
 # ]
 
 def loginPage(request):
+    if request.user.is_authenticated:
+        return redirect("home")
     if request.method=="POST":
         username = request.POST.get("username")
         password = request.POST.get("password")
@@ -66,6 +68,7 @@ def room(request, pk):
     context={"room": rooms}
     return render(request, "base/room.html", context)
 
+@login_required(login_url="login")
 def createRoom(request):
     form = RoomForm()
     if request.method == "POST":
@@ -77,9 +80,12 @@ def createRoom(request):
     context = {"form": form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url="login")
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here")
     if request.method=="POST":
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -88,8 +94,11 @@ def updateRoom(request, pk):
     context = {"form": form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url="login")
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here")
     if request.method == "POST":
         room.delete()
         return redirect("home")
